Add System property in VM to get omap enhancement flag
Allows an application to detect for an omap enhanced device
with a System.getProperty() call. Application should
check for for the omap.enhancement property to be true.
i.e "true".equals(System.getProperty("omap.enhancement"))

getProperty will return null string if OMAP_ENHANCEMENT is
not defined and result of comparison will be false.

Change-Id: Iae1ae4f6f7cc36fe2d59fdbfe0c5be4ee085d585