Add System property in VM to get omap enhancement flag
Allows an application to detect for an omap enhanced device
with a System.getProperty() call. Application should
check for for the omap.enhancement property to be true.
i.e "true".equals(System.getProperty("omap.enhancement"))

getProperty will return null string if OMAP_ENHANCEMENT is
not defined and result of comparison will be false.

Change-Id: Ie167b499f63e17d02b6bd3784d09a5424306e543